Allow trigger suggest without typing \
This could be useful because typing a \ might be harder/not as ergonomic
than pressing ctrl+space for some people.

I essentially added a branch checking for the CompletionTriggerKind.
And if it is (manual) Invoke, it does not assume that there is a \ to
begin with, but inserts it using a additionalTextEdit if necessary.

Right now it works when pressing space after

- a partial command (`\a|`), like before
- a new word (after a space/word boundary) and typing at least one character

It bugs when used after

- a new word and just using arrow keys to select a suggestion
	-> missing \
- an already present word, i.e. a partial command without leading \
	-> missing \ and not replacing the word

(I also partially refactored the code a bit in the process and hope it is more
readable now.)