Updated Anchor Tag to allow for Link Reference Decleratio
No parenthesis and optioanl white space.

Also updated to allow archor after link (link#anchor) but the anchors listed are local not from the linked document which is not correct. Fixing this is tricky, a tempory fix may be to simply not list archors from other files.

Need to apply same logic to File Paths autocomplete next