removed ompletion of headers from links for now
link url is captured but link is not opened or TOC generaed for archors can not be listed in autocomplete.

I'm unsure how to correctly parse in a sepereate document, but it seems like changing document to documentFromUrl on line 564 (where that is an avaliable TextDocument with .md extension) everything else would just work. But as i dont know how to do that then displaying nothing is better.