Better latex regex for completion
Previous capture relied on counting number of backslashes after matching a regex and performing modular arithmetic to check for an odd number.

This commit is to change the latexCmd regex to a single line which says:
`/(^|[^\\])([\\]{2})*\\$/`

```
IF ( ( (on new line) OR (following a non-backslash char) )
   AND (0 or more pairs of backslashes)
   AND a single backslash )
THEN there are an odd number of backslashes.
```

The benefit of this is to better align with the other regex's used in the if-else statement. also removes need for `matches` variable.